+++ docs/changelog.rst
@@ -2,7 +2,7 @@
 Changelog
 =========
 
-* :feature:`92` Format multiline locals better
+* :feature:`92` Improve indentation of multiline locals
 * :support:`90` Added support for python 3.7
 * :support:`88` Document how to install with conda
 * :release:`0.9.5 <2018-06-24>`